requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
dj-database-url==2.0.0
Django==4.2.3
django-crispy-forms==2.0
django-debug-toolbar==4.1.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
whitenoise==6.5.0
